Modify Register Screen
- /backend/: 바코드의 code 타입 Double로변경
- assets/: 이미지 추가
- src/common/images.ts: 이미지 경로 추가
- src/components/Common/Footer.tsx: 등록화면 라우팅 파라미터 값 추가
- src/components/Register/DatePicker.tsx: 날짜 선택 컴포넌트
- src/components/index.ts: 컴포넌트 경로 추가
- src/navigations/BottomTab.tsx: 파라미터 저장
- src/navigations/MainStack.tsx: 주석 수정
- src/screens/Register/BarcodeScreen.tsx: 바코드 등록 화면
- src/screens/Register/RegisterScreen.tsx: 상품 등록 화면 구현
- src/screens/index.ts: Screen 파일 경로 업데이트


Former-commit-id: 87546145ff75c7cd03769174c9e48069c66f889c

+++ backend/src/main/java/com/ssafy/eurekka/service/BarcodeService.java
@@ -4,5 +4,5 @@ import com.ssafy.eurekka.vo.Barcode;
 
 public interface BarcodeService {
 
-  Barcode findBarcode(int code);
+  Barcode findBarcode(Double code);
 }
